+++ app/src/main/java/com/zikey/android/razancatalogapp/model/ProductSubGroup.kt
@@ -3,7 +3,7 @@ package com.zikey.android.razancatalogapp.model
 import com.google.gson.annotations.SerializedName
 
 data class ProductSubGroup(
-    @SerializedName("i") val id: Int?,
+    @SerializedName("i") val id: Long?,
     @SerializedName("n") var name: String?,
     @SerializedName("p") var products: List<Product>?,
     @SerializedName("iu") val imageURL1: String?,
